Added UISegmentedControl as game mode switch.
Added two-segment UISegmentedControl to act as the game mode selector. The
segments have title values of `2` and `3` respectively. Using the
`titleForSegmentAtIndex` method on the UISegmentControl and converting the
result to an integer, setting a new property on the CardMatchingGame:
numberOfCardsToMatch to either 2 or 3.

Need to implement new matching logic to account for 2 or 3 card matching.